refactor: break index.js into 3 files, move to src
