+++ src/client/logging.js
