+++ src/server/index.js
@@ -1,25 +1,18 @@
 const osc = require("osc");
 const express = require("express");
 const WebSocket = require("ws");
-const os = require("os");
 const marked = require("marked");
 const readFileSync = require("fs").readFileSync;
-const {
-  DescribeTableCommand,
-  DynamoDBClient,
-} = require("@aws-sdk/client-dynamodb");
-const { SSMClient, GetParameterCommand } = require("@aws-sdk/client-ssm");
-const {
-  ScanCommand,
-  PutCommand,
-  DynamoDBDocumentClient,
-} = require("@aws-sdk/lib-dynamodb");
-const { fromSSO } = require("@aws-sdk/credential-provider-sso");
-const { fromInstanceMetadata } = require("@aws-sdk/credential-providers");
 const { uniqueNamesGenerator, names } = require("unique-names-generator");
 const ipInt = require("ip-to-int");
+const generateNameForIp = require("./utils").generateNameForIp;
+const getAWSCredentialsDependingOnEnvironment =
+  require("./utils").getAWSCredentialsDependingOnEnvironment;
+const saveControlMessage = require("./ddb").saveControlMessage;
+const getControlMessages = require("./ddb").getControlMessages;
 
 // express
+
 const app = express();
 const port = 8080;
 server = app.listen(port, () => {
@@ -29,22 +22,21 @@ app.use(express.static("dist"));
 app.use(express.json());
 
 app.get("/", (req, res) => {
-  res.sendFile(__dirname + "/dist/index.html");
+  res.sendFile(__dirname + "../dist/index.html");
 });
 
 app.get("/spec", (req, res) => {
-  var path = __dirname + "/doc/spec.md";
+  var path = __dirname + "../doc/spec.md";
   var file = readFileSync(path, "utf8");
   res.send(marked.parse(file.toString()));
 });
 
 app.get("/api", (req, res) => {
-  var path = __dirname + "/doc/api.md";
+  var path = __dirname + "../doc/api.md";
   var file = readFileSync(path, "utf8");
   res.send(marked.parse(file.toString()));
 });
 
-// todo: fix printing the OSC message in my log lines
 app.post("/api/send-message", (req, res) => {
   console.log(
     `Received OSC message via API: ${req.body}, redirecting it to UDP port`,
@@ -63,68 +55,14 @@ app.post("/api/send-message", (req, res) => {
   );
 });
 
-var credentials;
-if (process.env.NODE_ENV == "production") {
-  credentials = fromInstanceMetadata()();
-} else {
-  credentials = fromSSO({ profile: "osc-chat" })();
-}
-
-const ddbClient = new DynamoDBClient({
-  region: "eu-west-2",
-  credentials: credentials,
-});
-const docClient = DynamoDBDocumentClient.from(ddbClient);
-const tableName = (async () => {
-  const ssmClient = new SSMClient({
-    region: "eu-west-2",
-    credentials: credentials,
-  });
-  const command = new GetParameterCommand({
-    Name: "/osc-chat/dynamodb-table-name",
-  });
-  const response = await ssmClient.send(command);
-  console.log(`Retrieved DynamoDB table name: ${response.Parameter.Value}`);
-  return response.Parameter.Value;
-})();
-
-const saveControlMessage = async (oscMsg) => {
-  console.log(
-    `Saving control message ${JSON.stringify(oscMsg)} to DynamoDB table: ${await tableName}`,
-  );
-
-  const command = new PutCommand({
-    TableName: await tableName,
-    Item: {
-      channelAndGroup: `${oscMsg.args[1][0]}${oscMsg.args[1][1]}`,
-      channel: oscMsg.address,
-      args: oscMsg.args,
-      timestamp: Date.now().toString(),
-    },
-  });
-
-  const response = await docClient.send(command);
-  return response;
-};
-
 app.get("/api/get-control-messages", async (req, res) => {
-  console.log(
-    `Fetching all control messages from DynamoDB table: ${await tableName}`,
-  );
-
-  const command = new ScanCommand({
-    TableName: await tableName,
-  });
-
-  const response = await docClient.send(command);
-  const messages = response.Items.map((item) => ({
-    address: item.channel,
-    args: ["loader", item.args],
-  }));
+  const messages = await getControlMessages();
   console.log(`Retrieved control messages: ${JSON.stringify(messages)}`);
   res.setHeader("Content-Type", "application/json");
   res.end(JSON.stringify({ controlMessages: messages }));
 });
+
+// OSC UDP Port
 
 const udpPort = new osc.UDPPort({
   localAddress: "0.0.0.0",
@@ -133,6 +71,7 @@ const udpPort = new osc.UDPPort({
   remotePort: 8081,
   broadcast: true,
 });
+
 console.log("UDP port created on 0.0.0.0:57121");
 
 udpPort.on("message", (oscMsg, timeTag, info) => {
@@ -152,6 +91,7 @@ udpPort.on("ready", () => {
 udpPort.open();
 
 // WebSocket Server
+
 const wss = new WebSocket.Server({
   port: 8081,
 });
@@ -166,15 +106,16 @@ wss.on("connection", (socket) => {
     console.log(
       `Received OSC message via UDP: ${JSON.stringify(oscMsg)}, redirecting it to WebSocket.`,
     );
-    const seed = ipInt(info.address).toInt();
-    const name = uniqueNamesGenerator({
-      dictionaries: [names],
-      seed: seed,
-    }).toLowerCase();
     try {
       socketPort.send({
         address: oscMsg.address,
-        args: [{ type: "s", value: name }, oscMsg.args],
+        args: [
+          {
+            type: "s",
+            value: generateNameForIp(info.address),
+          },
+          oscMsg.args,
+        ],
       });
     } catch (error) {
       if (error.code === "ERR_UNHANDLED_ERROR") {
@@ -190,10 +131,6 @@ wss.on("connection", (socket) => {
   }
 
   udpPort.on("message", forwardMessageToWebSocket);
-
-  // var relay = new osc.Relay(udpPort, socketPort, {
-  //   raw: true,
-  // });
 
   socketPort.on("message", (oscMsg) => {
     console.log(
