+++ src/server/utils.js
@@ -1,0 +1,25 @@
+const { uniqueNamesGenerator, names } = require("unique-names-generator");
+const ipInt = require("ip-to-int");
+const { fromSSO } = require("@aws-sdk/credential-provider-sso");
+const { fromInstanceMetadata } = require("@aws-sdk/credential-providers");
+
+const generateNameForIp = (ip) => {
+  const seed = ipInt(ip).toInt();
+  return uniqueNamesGenerator({
+    dictionaries: [names],
+    seed: seed,
+  }).toLowerCase();
+};
+
+const getAWSCredentialsDependingOnEnvironment = () => {
+  if (process.env.NODE_ENV == "production") {
+    return fromInstanceMetadata()();
+  } else {
+    return fromSSO({ profile: "osc-chat" })();
+  }
+};
+
+module.exports = {
+  generateNameForIp,
+  getAWSCredentialsDependingOnEnvironment,
+};
